refac : profile refactor
